Create MealTrain app with models, serializers, views, and admin registration; update settings and URLs
