+++ backend/backend/urls.py
@@ -20,4 +20,5 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path("api/", include("user.urls")),
+    path("api/", include("mealTrain.urls")),
 ]
