+++ backend/mealTrain/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import MealTrain, MealTrainParticipant
+
+admin.site.register(MealTrain)
+admin.site.register(MealTrainParticipant)
