+++ backend/mealTrain/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MealtrainConfig(AppConfig):
+    name = 'mealTrain'
