+++ backend/user/admin.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
 
-from .models import MealTrain, MealTrainParticipant, Profile
+from .models import Profile
 
 admin.site.register(Profile)
-admin.site.register(MealTrain)
-admin.site.register(MealTrainParticipant)
