+++ backend/user/serializers.py
@@ -3,9 +3,6 @@ from django.contrib.auth.password_validation import validate_password
 from rest_framework import serializers
 from rest_framework_simplejwt.tokens import RefreshToken
 
-from .models import MealTrain
-from .models import Profile, MealTrain, MealTrainParticipant
-
 User = get_user_model()
 
 
@@ -79,30 +76,6 @@ class MeSerializer(serializers.ModelSerializer):
             "bio": p.bio,
             "allergies": p.allergies,
         }
-
-
-class MealTrainSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = MealTrain
-        fields = (
-            "id",
-            "title",
-            "description",
-            "start_date",
-            "end_date",
-            "restrictions",
-            "creator",
-            "created_at",
-            "updated_at",       
-        )
-        read_only_fields = ("id", "creator", "created_at", "updated_at")
-
-    def validate(self, attrs):
-        start = attrs.get("start_date")
-        end = attrs.get("end_date")
-        if start and end and start > end:
-            raise serializers.ValidationError("Start date must be before end date.")
-        return attrs
 
 
 def tokens_for_user(user) -> dict:
